# src/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "descriptuon": "sample server",
        "version": "1.0.0",
        "title": "Assignment Test",
        "contact": {
            "email": "arun.ang.js@gmail.com"
        }
    },
    "host": "localhost:3000",
    "basePath": "/api",
...
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/user": {
            "post": {
                "tags": ["User"],
                "Summary": "Create a user data",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "0"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Unable to create a new user !!"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "1"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Created a new user !!"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": ["User"],
                "Summary": "Get a user data",
                "parameters": [],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "1"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Successfully get the user list !!"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "put": {
                "tags": ["User"],
                "Summary": "Update a user data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "0"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Unable to update the user !!"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "1"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Successfully update the policy !!"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["User"],
                "Summary": "Delete a policy data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "0"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Invalid policy id !!"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "1"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Successfully delete a policy !!"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user-account": {
            "post": {
                "tags": ["User Account"],
                "Summary": "Create a user account data",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userAccount"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "0"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Unable to create a new user account !!"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "1"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Created a new user account !!"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": ["User Account"],
                "Summary": "Get a user account data",
                "parameters": [],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "1"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Successfully get the user account list !!"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user-account/{id}": {
            "put": {
                "tags": ["User Account"],
                "Summary": "Update a user data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "0"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Unable to update the user account!!"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "1"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Successfully update the User account !!"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["User Account"],
                "Summary": "Delete a user account data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "0"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Invalid user account id !!"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "1"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Successfully delete a User account !!"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "user": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "dob": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                }
            }
        },
        "userAccount": {
            "type": "object",
            "properties": {
                "userId": {
                    "type": "string"
                },
                "accountNumber": {
                    "type": "string"
                },
                "balance": {
                    "type": "string"
                }
            }
        }        
    }
}